requirements.txt
discord.py
discord-ext-voice-recv @ git+https://github.com/imayhaveborkedit/discord-ext-voice-recv
numpy
faster-whisper
torch
torchaudio
psutil

python-dotenv
